requirements.txt
sympy~=1.12
pywebview~=5.3.2
ttkbootstrap~=1.10.1
